# (PC-12384) Improve performance on adage routes by removing join to mediations (and therefore product used for this)