--- a/api/src/pcapi/routes/adage/v1/serialization/prebooking.py
+++ b/api/src/pcapi/routes/adage/v1/serialization/prebooking.py
@@ -122,7 +122,7 @@ def serialize_educational_booking(educational_booking: EducationalBooking) -> Ed
         durationMinutes=offer.durationMinutes,
         expirationDate=booking.expirationDate,
         id=educational_booking.id,
-        image={"url": offer.image.url, "credit": offer.image.credit} if offer.image else None,
+        image=None,
         isDigital=offer.isDigital,
         venueName=venue.name,
         name=offer.name,