--- a/api/tests/routes/adage/v1/get_educational_institution_test.py
+++ b/api/tests/routes/adage/v1/get_educational_institution_test.py
@@ -88,7 +88,7 @@ class Returns200Test:
                     "durationMinutes": offer.durationMinutes,
                     "expirationDate": booking.expirationDate,
                     "id": educational_booking.id,
-                    "image": {"url": offer.image.url, "credit": offer.image.credit} if offer.image else None,
+                    "image": None,
                     "isDigital": offer.isDigital,
                     "venueName": venue.name,
                     "name": offer.name,