Lien vers le ticket Jira : https://passculture.atlassian.net/browse/PC-12384

**Avant:**
```
 Gather  (cost=869049.33..944794.95 rows=44165 width=4995) (actual time=217.726..234.458 rows=4 loops=1)
   Workers Planned: 1
   Workers Launched: 1
   ->  Parallel Hash Left Join  (cost=868049.33..939378.45 rows=25979 width=4995) (actual time=211.707..212.599 rows=2 loops=2)
         Hash Cond: (offer_1.id = mediation_1."offerId")
         ->  Hash Join  (cost=855115.93..893149.89 rows=25979 width=4922) (actual time=31.446..32.046 rows=2 loops=2)
               Hash Cond: (venue_1.id = offer_1."venueId")
               ->  Parallel Seq Scan on venue venue_1  (cost=0.00..1455.07 rows=16307 width=628) (actual time=0.016..8.160 rows=14524 loops=2)
               ->  Hash  (cost=831272.87..831272.87 rows=44165 width=4294) (actual time=1.275..1.286 rows=4 loops=2)
                     Buckets: 1024  Batches: 64  Memory Usage: 11kB
                     ->  Nested Loop  (cost=2.58..831272.87 rows=44165 width=4294) (actual time=1.037..1.155 rows=4 loops=2)
                           ->  Nested Loop  (cost=2.15..561032.21 rows=44165 width=2946) (actual time=0.190..0.293 rows=4 loops=2)
                                 ->  Merge Join  (cost=1.59..280926.76 rows=44165 width=1254) (actual time=0.171..0.254 rows=4 loops=2)
                                       Merge Cond: (educational_booking."educationalRedactorId" = educational_redactor.id)
                                       ->  Nested Loop  (cost=1.44..280108.98 rows=44165 width=382) (actual time=0.154..0.231 rows=4 loops=2)
                                             ->  Nested Loop  (cost=0.88..341.81 rows=44165 width=252) (actual time=0.137..0.195 rows=4 loops=2)
                                                   ->  Nested Loop  (cost=0.44..325.34 rows=1 width=128) (actual time=0.123..0.163 rows=4 loops=2)
                                                         ->  Nested Loop  (cost=0.29..317.16 rows=1 width=128) (actual time=0.107..0.132 rows=4 loops=2)
                                                               Join Filter: (educational_booking."educationalInstitutionId" = educational_institution.id)
                                                               Rows Removed by Join Filter: 109
                                                               ->  Index Scan using "ix_educational_booking_educationalRedactorId" on educational_booking  (cost=0.14..306.05 rows=196 width=46) (actual time=0.011..0.054 rows=113 loops=2)
                                                                     Filter: (("educationalYearId")::text = '7'::text)
                                                               ->  Materialize  (cost=0.15..8.17 rows=1 width=82) (actual time=0.000..0.000 rows=1 loops=226)
                                                                     ->  Index Scan using "ix_educational_institution_institutionId" on educational_institution  (cost=0.15..8.17 rows=1 width=82) (actual time=0.017..0.018 rows=1 loops=2)
                                                                           Index Cond: (("institutionId")::text = '0220186H'::text)
                                                         ->  Index Only Scan using "educational_year_adageId_key" on educational_year  (cost=0.15..8.17 rows=1 width=78) (actual time=0.006..0.006 rows=1 loops=8)
                                                               Index Cond: ("adageId" = '7'::text)
                                                               Heap Fetches: 8
                                                   ->  Index Scan using "ix_booking_educationalBookingId" on booking booking_1  (cost=0.43..8.45 rows=1 width=124) (actual time=0.005..0.005 rows=1 loops=8)
                                                         Index Cond: ("educationalBookingId" = educational_booking.id)
                                             ->  Index Scan using stock_pkey on stock stock_1  (cost=0.56..6.33 rows=1 width=130) (actual time=0.006..0.006 rows=1 loops=8)
                                                   Index Cond: (id = booking_1."stockId")
                                       ->  Index Scan using educational_redactor_pkey on educational_redactor  (cost=0.14..265.49 rows=90 width=872) (actual time=0.005..0.009 rows=25 loops=2)
                                 ->  Index Scan using offer_pkey on offer offer_1  (cost=0.56..6.34 rows=1 width=1692) (actual time=0.007..0.007 rows=1 loops=8)
                                       Index Cond: (id = stock_1."offerId")
                           ->  Index Scan using product_pkey on product product_1  (cost=0.43..6.12 rows=1 width=1348) (actual time=0.212..0.212 rows=1 loops=8)
                                 Index Cond: (id = offer_1."productId")
         ->  Parallel Hash  (cost=9387.29..9387.29 rows=140729 width=73) (actual time=139.537..139.538 rows=169266 loops=2)
               Buckets: 65536  Batches: 16  Memory Usage: 2944kB
               ->  Parallel Seq Scan on mediation mediation_1  (cost=0.00..9387.29 rows=140729 width=73) (actual time=0.011..51.655 rows=169266 loops=2)
 Planning Time: 3.848 ms
 Execution Time: 234.871 ms
```

**Après:**
```
                                                                                                     QUERY PLAN
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Hash Join  (cost=4149.78..579692.12 rows=44165 width=3574) (actual time=47.401..47.956 rows=4 loops=1)
   Hash Cond: (educational_booking."educationalRedactorId" = educational_redactor.id)
   ->  Hash Join  (cost=4137.76..579558.70 rows=44165 width=2702) (actual time=47.345..47.893 rows=4 loops=1)
         Hash Cond: (offer_1."venueId" = venue_1.id)
         ->  Nested Loop  (cost=2.01..550399.00 rows=44165 width=2074) (actual time=0.180..0.304 rows=4 loops=1)
               ->  Nested Loop  (cost=1.44..275183.27 rows=44165 width=382) (actual time=0.153..0.255 rows=4 loops=1)
                     ->  Nested Loop  (cost=0.88..341.81 rows=44165 width=252) (actual time=0.138..0.218 rows=4 loops=1)
                           ->  Nested Loop  (cost=0.44..325.34 rows=1 width=128) (actual time=0.125..0.181 rows=4 loops=1)
                                 ->  Nested Loop  (cost=0.29..317.16 rows=1 width=128) (actual time=0.095..0.133 rows=4 loops=1)
                                       Join Filter: (educational_booking."educationalInstitutionId" = educational_institution.id)
                                       Rows Removed by Join Filter: 109
                                       ->  Index Scan using "ix_educational_booking_educationalRedactorId" on educational_booking  (cost=0.14..306.05 rows=196 width=46) (actual time=0.008..0.066 rows=113 loops=1)
                                             Filter: (("educationalYearId")::text = '7'::text)
                                       ->  Materialize  (cost=0.15..8.17 rows=1 width=82) (actual time=0.000..0.000 rows=1 loops=113)
                                             ->  Index Scan using "ix_educational_institution_institutionId" on educational_institution  (cost=0.15..8.17 rows=1 width=82) (actual time=0.010..0.011 rows=1 loops=1)
                                                   Index Cond: (("institutionId")::text = '0220186H'::text)
                                 ->  Index Only Scan using "educational_year_adageId_key" on educational_year  (cost=0.15..8.17 rows=1 width=78) (actual time=0.009..0.011 rows=1 loops=4)
                                       Index Cond: ("adageId" = '7'::text)
                                       Heap Fetches: 4
                           ->  Index Scan using "ix_booking_educationalBookingId" on booking booking_1  (cost=0.43..8.45 rows=1 width=124) (actual time=0.007..0.007 rows=1 loops=4)
                                 Index Cond: ("educationalBookingId" = educational_booking.id)
                     ->  Index Scan using stock_pkey on stock stock_1  (cost=0.56..6.22 rows=1 width=130) (actual time=0.007..0.007 rows=1 loops=4)
                           Index Cond: (id = booking_1."stockId")
               ->  Index Scan using offer_pkey on offer offer_1  (cost=0.56..6.23 rows=1 width=1692) (actual time=0.010..0.010 rows=1 loops=4)
                     Index Cond: (id = stock_1."offerId")
         ->  Hash  (cost=1569.22..1569.22 rows=27722 width=628) (actual time=43.868..43.869 rows=29048 loops=1)
               Buckets: 8192  Batches: 8  Memory Usage: 911kB
               ->  Seq Scan on venue venue_1  (cost=0.00..1569.22 rows=27722 width=628) (actual time=0.015..15.936 rows=29048 loops=1)
   ->  Hash  (cost=10.90..10.90 rows=90 width=872) (actual time=0.017..0.017 rows=33 loops=1)
         Buckets: 1024  Batches: 1  Memory Usage: 11kB
         ->  Seq Scan on educational_redactor  (cost=0.00..10.90 rows=90 width=872) (actual time=0.008..0.010 rows=33 loops=1)
 Planning Time: 6.347 ms
 Execution Time: 48.290 ms
(33 rows)
```